% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neostocks.R
\docType{package}
\name{neostocks-package}
\alias{neostocks}
\alias{neostocks-package}
\title{neostocks: Real-time stock tracker for the Neopets Stock Market}
\description{
Real-time stock tracker for the Neopets Stock Market.
}
\author{
\strong{Maintainer}: Greg Lin \email{glin@glin.io}

}
\keyword{internal}
