% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{new_app}
\alias{new_app}
\title{Create a new app}
\usage{
new_app()
}
\description{
Uses the following environment variables:
 * NEOSTOCKS_DATA_FILE   path to the stock data file or directory (defaults to example data)
 * NEOSTOCKS_UI_DIR      path to the UI directory (defaults to `dist`)
 * NEOSTOCKS_CACHE_DIR   path to the cache directory (optional)
 * NEOSTOCKS_ENABLE_API  if `true` (default), enable the internal API
}
