% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{new_app}
\alias{new_app}
\title{Create a new app}
\usage{
new_app()
}
\description{
Uses the following environment variables:
 * NEOSTOCKS_DATA_FILE   path to the stock data file
 * NEOSTOCKS_UI_DIR      path to the UI directory
 * NEOSTOCKS_ENABLE_API  if `true`, enable the internal API
}
